+++ AndroidApps/app/src/main/AndroidManifest.xml
@@ -24,6 +24,9 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
+            android:name=".view.activity.LocationActivity"
+            android:exported="false" />
+        <activity
             android:name=".view.custom.CustomAlertDialog"
             android:exported="false" />
         <activity
