+++ AndroidApps/app/src/main/java/com/mfa/api/retrofit/ApiConfig.kt
@@ -16,7 +16,7 @@ object ApiConfig {
             .build()
         val retrofit = Retrofit.Builder()
 //            .baseUrl("http://192.168.0.107:3000/")
-            .baseUrl("http://192.168.0.244:3000/")
+            .baseUrl("http://192.168.0.221:3000/")
             .addConverterFactory(GsonConverterFactory.create())
             .client(client)
             .build()
