+++ AndroidApps/app/src/main/java/com/mfa/view/activity/HomeActivity.kt
@@ -51,6 +51,8 @@ class HomeActivity : AppCompatActivity() {
         const val NAME = "name"
     }
 
+
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityHomeBinding.inflate(layoutInflater)
