+++ AndroidApps/app/src/main/java/com/mfa/view/adapter/PertemuanAdapter.kt
@@ -12,6 +12,7 @@ import com.mfa.api.response.PertemuanResponseItem
 import com.mfa.databinding.ItemPertemuanBinding
 import com.mfa.view.activity.PertemuanActivity
 import com.mfa.view.activity.PresensiActivity
+import com.mfa.view.adapter.JadwalAdapter.Companion.DIFF_CALLBACK
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
@@ -35,8 +36,9 @@ class PertemuanAdapter :
             holder.itemView.setOnClickListener {
                 val sendData = Intent(holder.itemView.context, PresensiActivity::class.java)
                 Log.d("PertemuanAdapter", "Sending jadwal: ${pertemuan.jadwal}, ruang: ${pertemuan.jadwal}")
-                sendData.putExtra(PresensiActivity.GETJADWAL, pertemuan.jadwal)
+                sendData.putExtra(PresensiActivity.GETJADWAL, "${pertemuan.jadwal}  grup ${pertemuan.grup}")
                 sendData.putExtra(PresensiActivity.RUANG, pertemuan.ruang)
+//                sendData.putExtra(PresensiActivity.NAMAPERTEMUAN, "${pertemuan.mataKuliah}")
                 sendData.putExtra(PresensiActivity.ISVALID,isvalid)
                 holder.itemView.context.startActivity(sendData)
             }
