+++ AndroidApps/app/src/main/res/drawable/bg_button.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <!-- Shadow -->
+    <item>
+        <shape android:shape="rectangle">
+            <corners android:radius="12dp" />
+            <solid android:color="@color/green_light" />
+            <padding android:top="4dp" />
+        </shape>
+    </item>
+
+    <!-- Button background -->
+    <item android:top="2dp">
+        <shape android:shape="rectangle">
+            <corners android:radius="12dp" />
+            <solid android:color="@color/green_primary" />
+        </shape>
+    </item>
+</layer-list>
