+++ AndroidApps/app/src/main/res/drawable/dialog_background.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <corners android:radius="16dp"/>
+    <solid android:color="@android:color/white"/>
+    <stroke android:width="2dp" android:color="@color/teal_700"/>
+</shape>
