+++ AndroidApps/app/src/main/res/layout/activity_home.xml
@@ -12,7 +12,7 @@
         android:id="@+id/topAppBar"
         android:layout_width="match_parent"
         android:layout_height="?attr/actionBarSize"
-        android:background="@color/teal_700"
+        android:background="@color/green_primary"
         android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
         app:title="AbzennDW"
         app:titleTextColor="@android:color/white"
@@ -99,7 +99,7 @@
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:background="@color/teal_700"
+            android:background="@color/green_primary"
             android:orientation="vertical"
             android:padding="16dp">
