+++ AndroidApps/app/src/main/res/layout/activity_pertemuan.xml
@@ -12,7 +12,7 @@
         android:id="@+id/topAppBar"
         android:layout_width="match_parent"
         android:layout_height="?attr/actionBarSize"
-        android:background="@color/teal_700"
+        android:background="@color/green_primary"
         android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
         app:title="RPLBO Grup A"
         app:titleTextColor="@android:color/white"
