+++ AndroidApps/app/src/main/res/layout/activity_presensi.xml
@@ -10,54 +10,121 @@
         android:layout_height="wrap_content"
         tools:context=".view.PresensiActivityActivity">
 
+        <androidx.appcompat.widget.Toolbar
+            android:id="@+id/topAppBar"
+            android:layout_width="match_parent"
+            android:layout_height="?attr/actionBarSize"
+            android:background="@color/green_primary"
+            android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
+            app:title="RPLBO Grup A"
+            app:titleTextColor="@android:color/white"
+            app:navigationIcon="@drawable/arrow_left"
+            app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
+            app:layout_constraintTop_toTopOf="parent" />
+
 
         <androidx.cardview.widget.CardView
             android:id="@+id/cardView"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:layout_marginTop="5dp"
-
+            android:layout_marginTop="60dp"
+            android:layout_marginStart="20dp"
+            android:layout_marginEnd="20dp"
             android:layout_marginBottom="10dp"
+            app:cardCornerRadius="12dp"
+            app:cardElevation="4dp"
+            app:cardBackgroundColor="@color/green_primary"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
+            app:layout_constraintTop_toBottomOf="@id/topAppBar">
 
             <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:background="@drawable/bg_new"
+                android:background="@color/green_primary"
                 android:orientation="vertical"
-                android:padding="16dp">
+                android:padding="30dp">
 
                 <TextView
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="@string/position"
-                    android:textSize="24sp"
-                    android:id="@+id/position"
+                    android:text="RPLBO pertemuan 4"
+                    android:textSize="20sp"
+                    android:layout_marginBottom="10dp"
+                    android:textStyle="bold"
                     android:textColor="@color/white"/>
 
                 <TextView
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="@string/status"
-                    android:textSize="24sp"
-                    android:id="@+id/status"
+                    android:text="Waktu      : 11:30 - 14:20"
+                    android:textSize="16sp"
+                    android:layout_marginBottom="5dp"
                     android:textColor="@color/white"/>
+
+                <TextView
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="Ruangan  : Lab H"
+                    android:textSize="16sp"
+                    android:layout_marginBottom="5dp"
+                    android:textColor="@color/white"/>
+
+                <TextView
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="Tanggal   : 25 March 2024"
+                    android:textSize="16sp"
+                    android:textColor="@color/white"
+                    android:layout_marginBottom="5dp"/>
+
+                <TextView
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="Dosen      : Dendy Senapartha"
+                    android:textSize="16sp"
+                    android:layout_marginBottom="5dp"
+                    android:textColor="@color/white"/>
+
+                <LinearLayout
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:orientation="horizontal">
+
+                    <TextView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:text="Status      : "
+                        android:textSize="16sp"
+                        android:textColor="@color/white"/>
+
+                    <TextView
+                        android:id="@+id/status"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:text="Alpha"
+                        android:textSize="16sp"
+                        android:textStyle="bold"
+                        android:textColor="@color/red"/>
+                </LinearLayout>
 
             </LinearLayout>
 
         </androidx.cardview.widget.CardView>
 
+
         <Button
-            android:id="@+id/btnScanQr"
+            android:id="@+id/btnPresensi"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:layout_marginTop="500dp"
+            android:layout_marginTop="250dp"
+            android:layout_marginStart="60dp"
+            android:layout_marginEnd="60dp"
             android:paddingEnd="15dp"
             android:paddingStart="15dp"
             android:padding="15dp"
-            android:text="@string/scan_qr"
+            android:background="@drawable/bg_button"
+            android:text="Presensi"
             app:layout_constraintStart_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@id/cardView" />
 
@@ -65,10 +132,13 @@
             android:id="@+id/btnBackHome"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:layout_marginTop="500dp"
+            android:layout_marginTop="250dp"
+            android:layout_marginStart="60dp"
+            android:layout_marginEnd="60dp"
             android:paddingEnd="15dp"
             android:paddingStart="15dp"
             android:padding="15dp"
+            android:background="@drawable/bg_button"
             android:visibility="gone"
             android:text="Kembali ke home"
             app:layout_constraintStart_toEndOf="parent"
@@ -84,15 +154,15 @@
             app:layout_constraintStart_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@id/cardView"/>
 
-        <ProgressBar
-            android:id="@+id/progressBar"
-            style="?android:attr/progressBarStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:layout_constraintBottom_toTopOf="@+id/btnScanQr"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/cardView" />
+<!--        <ProgressBar-->
+<!--            android:id="@+id/progressBar"-->
+<!--            style="?android:attr/progressBarStyle"-->
+<!--            android:layout_width="wrap_content"-->
+<!--            android:layout_height="wrap_content"-->
+<!--            app:layout_constraintBottom_toTopOf="@+id/btnPresensi"-->
+<!--            app:layout_constraintEnd_toEndOf="parent"-->
+<!--            app:layout_constraintStart_toStartOf="parent"-->
+<!--            app:layout_constraintTop_toBottomOf="@+id/cardView" />-->
 
 
     </androidx.constraintlayout.widget.ConstraintLayout>
