+++ AndroidApps/app/src/main/res/values/colors.xml
@@ -6,7 +6,7 @@
     <color name="teal_200">#FF03DAC5</color>
     <color name="teal_700">#FF018786</color>
     <color name="blue">#4286f4</color>
-    <color name="red">#EE1000</color>
+    <color name="red">#FF0000</color>
     <color name="blue_normal">#3C39D5</color>
     <color name="blue_dashboard">#2D2BA0</color>
